// src/user/lib/enums/pregnancy-milstones.enum.ts
export enum PregnancyMilestones {
  BABY_CONCEIVED = 'BABY_CONCEIVED',
  PREGNANCY_TEST_POSITIVE = 'PREGNANCY_TEST_POSITIVE',
  HEARTBEAT_DETECTABLE = 'HEARTBEAT_DETECTABLE',
  MISCARRIAGE_RISK_DECREASES = 'MISCARRIAGE_RISK_DECREASES',
  BABY_MOVEMENTS_DETECTED = 'BABY_MOVEMENTS_DETECTED',
  PREMATURE_SURVIVAL = 'PREMATURE_SURVIVAL',
  BABY_CAN_BREATHE = 'BABY_CAN_BREATHE',
  FULL_TERM = 'FULL_TERM',
  DEFAULT = 'DEFAULT',
}